WIP move predicted movement calculation to a separate plugin
