+++ simulator/autoware_dummy_perception_publisher/CMakeLists.txt
@@ -26,6 +26,7 @@ ament_auto_add_library(signed_distance_function SHARED
 ament_auto_add_executable(${PROJECT_NAME}_node
   src/main.cpp
   src/node.cpp
+  src/object_info.cpp
   src/pointcloud_creator.cpp
   src/predicted_object_movement_plugin.cpp
 )
