+++ simulator/autoware_dummy_perception_publisher/src/node.cpp
@@ -71,141 +71,6 @@ using geometry_msgs::msg::TwistWithCovariance;
 using tier4_simulation_msgs::msg::DummyObject;
 using InterpolationTrajectory = autoware::experimental::trajectory::Trajectory<Pose>;
 
-ObjectInfo::ObjectInfo(
-  const tier4_simulation_msgs::msg::DummyObject & object, const rclcpp::Time & current_time)
-: length(object.shape.dimensions.x),
-  width(object.shape.dimensions.y),
-  height(object.shape.dimensions.z),
-  std_dev_x(std::sqrt(object.initial_state.pose_covariance.covariance[0])),
-  std_dev_y(std::sqrt(object.initial_state.pose_covariance.covariance[7])),
-  std_dev_z(std::sqrt(object.initial_state.pose_covariance.covariance[14])),
-  std_dev_yaw(std::sqrt(object.initial_state.pose_covariance.covariance[35])),
-  twist_covariance_(object.initial_state.twist_covariance),
-  pose_covariance_(object.initial_state.pose_covariance)
-{
-  const auto current_pose = calculateStraightLinePosition(object, current_time);
-
-  // calculate tf from map to moved_object
-  Transform ros_map2moved_object;
-  ros_map2moved_object.translation.x = current_pose.position.x;
-  ros_map2moved_object.translation.y = current_pose.position.y;
-  ros_map2moved_object.translation.z = current_pose.position.z;
-  ros_map2moved_object.rotation = current_pose.orientation;
-  tf2::fromMsg(ros_map2moved_object, tf_map2moved_object);
-
-  // set twist and pose information
-  const double initial_vel = std::clamp(
-    object.initial_state.twist_covariance.twist.linear.x, static_cast<double>(object.min_velocity),
-    static_cast<double>(object.max_velocity));
-  const double initial_acc = object.initial_state.accel_covariance.accel.linear.x;
-  const double elapsed_time = current_time.seconds() - rclcpp::Time(object.header.stamp).seconds();
-  double current_vel = initial_vel + initial_acc * elapsed_time;
-  if (initial_acc != 0.0) {
-    current_vel = std::clamp(
-      current_vel, static_cast<double>(object.min_velocity),
-      static_cast<double>(object.max_velocity));
-  }
-
-  stopAtZeroVelocity(current_vel, initial_vel, initial_acc);
-
-  twist_covariance_.twist.linear.x = current_vel;
-  pose_covariance_.pose = current_pose;
-}
-
-ObjectInfo::ObjectInfo(
-  const tier4_simulation_msgs::msg::DummyObject & object, const PredictedObject & predicted_object,
-  const rclcpp::Time & predicted_time, const rclcpp::Time & current_time,
-  const double switch_time_threshold)
-: length(object.shape.dimensions.x),
-  width(object.shape.dimensions.y),
-  height(object.shape.dimensions.z),
-  std_dev_x(std::sqrt(object.initial_state.pose_covariance.covariance[0])),
-  std_dev_y(std::sqrt(object.initial_state.pose_covariance.covariance[7])),
-  std_dev_z(std::sqrt(object.initial_state.pose_covariance.covariance[14])),
-  std_dev_yaw(std::sqrt(object.initial_state.pose_covariance.covariance[35])),
-  twist_covariance_(object.initial_state.twist_covariance),
-  pose_covariance_(object.initial_state.pose_covariance)
-{
-  // Check if threshold time has passed since object creation
-  const double time_since_creation = (current_time - rclcpp::Time(object.header.stamp)).seconds();
-  // Use straight-line movement for first switch_time_threshold seconds, then switch to predicted
-  // path
-  if (
-    time_since_creation < switch_time_threshold ||
-    predicted_object.kinematics.predicted_paths.empty()) {
-    // Reuse the logic from the other constructor
-    *this = ObjectInfo(object, current_time);
-    return;
-  }
-
-  const auto interpolated_pose =
-    calculateTrajectoryBasedPosition(object, predicted_object, predicted_time, current_time);
-
-  // Update pose and transform
-  pose_covariance_.pose = interpolated_pose;
-  tf2::fromMsg(interpolated_pose, tf_map2moved_object);
-
-  // Use dummy object's velocity consistently
-  twist_covariance_.twist.linear.x = object.initial_state.twist_covariance.twist.linear.x;
-}
-
-void ObjectInfo::stopAtZeroVelocity(double & current_vel, double initial_vel, double initial_acc)
-{
-  // stop at zero velocity
-  if (initial_acc < 0 && initial_vel > 0) {
-    current_vel = std::max(current_vel, 0.0);
-  }
-  if (initial_acc > 0 && initial_vel < 0) {
-    current_vel = std::min(current_vel, 0.0);
-  }
-}
-
-Pose ObjectInfo::calculateStraightLinePosition(
-  const tier4_simulation_msgs::msg::DummyObject & object, const rclcpp::Time & current_time)
-{
-  const auto & initial_pose = object.initial_state.pose_covariance.pose;
-  const double initial_vel = std::clamp(
-    object.initial_state.twist_covariance.twist.linear.x, static_cast<double>(object.min_velocity),
-    static_cast<double>(object.max_velocity));
-  const double initial_acc = object.initial_state.accel_covariance.accel.linear.x;
-
-  const double elapsed_time = current_time.seconds() - rclcpp::Time(object.header.stamp).seconds();
-
-  double move_distance{0.0};
-  double current_vel = initial_vel + initial_acc * elapsed_time;
-  if (initial_acc == 0.0) {
-    move_distance = initial_vel * elapsed_time;
-    return autoware_utils_geometry::calc_offset_pose(initial_pose, move_distance, 0.0, 0.0);
-  }
-
-  stopAtZeroVelocity(current_vel, initial_vel, initial_acc);
-  // clamp velocity within min/max
-  current_vel = std::clamp(
-    current_vel, static_cast<double>(object.min_velocity),
-    static_cast<double>(object.max_velocity));
-
-  // add distance on acceleration or deceleration
-  move_distance = (std::pow(current_vel, 2) - std::pow(initial_vel, 2)) * 0.5 / initial_acc;
-
-  // add distance after reaching max_velocity
-  if (initial_acc > 0) {
-    const double time_to_reach_max_vel =
-      std::max(static_cast<double>(object.max_velocity) - initial_vel, 0.0) / initial_acc;
-    move_distance += static_cast<double>(object.max_velocity) *
-                     std::max(elapsed_time - time_to_reach_max_vel, 0.0);
-  }
-
-  // add distance after reaching min_velocity
-  if (initial_acc < 0) {
-    const double time_to_reach_min_vel =
-      std::min(static_cast<double>(object.min_velocity) - initial_vel, 0.0) / initial_acc;
-    move_distance += static_cast<double>(object.min_velocity) *
-                     std::max(elapsed_time - time_to_reach_min_vel, 0.0);
-  }
-
-  return autoware_utils_geometry::calc_offset_pose(initial_pose, move_distance, 0.0, 0.0);
-}
-
 Pose ObjectInfo::calculateTrajectoryBasedPosition(
   const tier4_simulation_msgs::msg::DummyObject & object, const PredictedObject & predicted_object,
   const rclcpp::Time & predicted_time, const rclcpp::Time & current_time)
@@ -406,63 +271,11 @@ void DummyPerceptionPublisherNode::timerCallback()
 
     // Try to find matching predicted object
     const auto & object = objects_.at(i);
-    auto predicted_object_pair = findMatchingPredictedObject(object.id, current_time);
-    const auto & predicted_object = predicted_object_pair.first;
-    const auto & predicted_time = predicted_object_pair.second;
-
-    const bool matched_predicted =
-      (!predicted_object.object_id.uuid.empty() &&
-       !predicted_object.kinematics.predicted_paths.empty() && predicted_time.nanoseconds() > 0);
-    auto & dummy_predicted_info_map =
-      predicted_dummy_objects_tracking_info_.dummy_predicted_info_map;
-
     ObjectInfo obj_info = [&]() {
-      if (object.action == tier4_simulation_msgs::msg::DummyObject::PREDICT) {
-        if (matched_predicted) {
-          return ObjectInfo(
-            object, predicted_object, predicted_time, current_time,
-            predicted_object_params_.switch_time_threshold);
-        }
-
-        // Check if we have a last used prediction for this object
-        const auto & dummy_uuid_str = autoware_utils_uuid::to_hex_string(object.id);
-
-        auto info_it = dummy_predicted_info_map.find(dummy_uuid_str);
-
-        if (
-          info_it != dummy_predicted_info_map.end() &&
-          info_it->second.last_used_prediction.has_value() &&
-          info_it->second.last_used_prediction_time.has_value()) {
-          RCLCPP_DEBUG(
-            rclcpp::get_logger("dummy_perception_publisher"),
-            "Using last known prediction for lost object with ID: %s", dummy_uuid_str.c_str());
-          return ObjectInfo(
-            object, info_it->second.last_used_prediction.value(),
-            info_it->second.last_used_prediction_time.value(), current_time,
-            predicted_object_params_.switch_time_threshold);
-        }
-
-        RCLCPP_DEBUG(
-          rclcpp::get_logger("dummy_perception_publisher"),
-          "No matching predicted object found for dummy object with ID: %s",
-          dummy_uuid_str.c_str());
-      }
-
       // Use straight-line motion (original constructor) for all other actions
       return ObjectInfo(object, current_time);
     }();
     obj_infos.push_back(obj_info);
-
-    // Update last known position based on calculated ObjectInfo position
-    const auto & dummy_uuid_str = autoware_utils_uuid::to_hex_string(object.id);
-    dummy_predicted_info_map[dummy_uuid_str].last_known_position =
-      obj_info.pose_covariance_.pose.position;
-
-    // Track object creation time if not already tracked
-    auto & info = dummy_predicted_info_map[dummy_uuid_str];
-    if (!info.creation_timestamp.has_value()) {
-      info.creation_timestamp = rclcpp::Time(object.header.stamp);
-    }
   }
 
   // publish ground truth
@@ -671,576 +484,4 @@ void DummyPerceptionPublisherNode::objectCallback(
   }
 }
 
-std::pair<PredictedObject, rclcpp::Time> DummyPerceptionPublisherNode::findMatchingPredictedObject(
-  const unique_identifier_msgs::msg::UUID & object_id, const rclcpp::Time & current_time)
-{
-  PredictedObject empty_object;
-  rclcpp::Time empty_time(0, 0, RCL_ROS_TIME);
-
-  const auto & obj_uuid_str = autoware_utils_uuid::to_hex_string(object_id);
-
-  // Check if this dummy object is mapped to a predicted object
-  auto & dummy_predicted_info_map = predicted_dummy_objects_tracking_info_.dummy_predicted_info_map;
-  auto mapping_it = dummy_predicted_info_map.find(obj_uuid_str);
-  if (mapping_it == dummy_predicted_info_map.end() || mapping_it->second.predicted_uuid.empty()) {
-    return std::make_pair(empty_object, empty_time);
-  }
-
-  const std::string & mapped_predicted_uuid = mapping_it->second.predicted_uuid;
-
-  // Check if we should keep using the current prediction for at least
-  // min_predicted_path_keep_duration Check if we have a last used prediction and if we should keep
-  // using it
-  auto info_it = dummy_predicted_info_map.find(obj_uuid_str);
-  if (info_it == dummy_predicted_info_map.end()) {
-    // Create entry if it doesn't exist (needed for later updates)
-    dummy_predicted_info_map[obj_uuid_str];
-    info_it = dummy_predicted_info_map.find(obj_uuid_str);
-  }
-
-  if (
-    info_it->second.last_used_prediction.has_value() &&
-    info_it->second.prediction_update_timestamp.has_value()) {
-    const double time_since_last_update =
-      (current_time - info_it->second.prediction_update_timestamp.value()).seconds();
-
-    // If less than min_predicted_path_keep_duration has passed since last update, keep using the
-    // same prediction
-    if (time_since_last_update < predicted_object_params_.min_predicted_path_keep_duration) {
-      if (info_it->second.last_used_prediction_time.has_value()) {
-        return std::make_pair(
-          info_it->second.last_used_prediction.value(),
-          info_it->second.last_used_prediction_time.value());
-      }
-    }
-  }
-
-  // Time to update: find the closest prediction in the past
-  const auto & predicted_objects_buffer =
-    predicted_dummy_objects_tracking_info_.predicted_objects_buffer;
-  for (auto it = predicted_objects_buffer.rbegin(); it != predicted_objects_buffer.rend(); ++it) {
-    const auto & predicted_objects_msg = *it;
-    const rclcpp::Time msg_time(predicted_objects_msg.header.stamp);
-
-    // Skip future messages
-    if (msg_time > current_time) {
-      continue;
-    }
-
-    // Look for the mapped predicted object UUID
-    for (const auto & predicted_object : predicted_objects_msg.objects) {
-      unique_identifier_msgs::msg::UUID pred_obj_uuid;
-      pred_obj_uuid.uuid = predicted_object.object_id.uuid;
-      const auto & pred_obj_uuid_str = autoware_utils_uuid::to_hex_string(pred_obj_uuid);
-
-      if (pred_obj_uuid_str == mapped_predicted_uuid) {
-        // Apply path selection strategy based on object type configuration
-        PredictedObject modified_predicted_object = predicted_object;
-
-        // Check if this is a pedestrian object
-        const bool is_pedestrian = std::any_of(
-          predicted_object.classification.begin(), predicted_object.classification.end(),
-          [](const auto & classification) {
-            return classification.label ==
-                   autoware_perception_msgs::msg::ObjectClassification::PEDESTRIAN;
-          });
-
-        // Determine path selection strategy based on object type
-        const auto & pedestrian_params = predicted_object_params_.pedestrian_params;
-        const auto & vehicle_params = predicted_object_params_.vehicle_params;
-        const std::string path_selection_strategy = is_pedestrian
-                                                      ? pedestrian_params.path_selection_strategy
-                                                      : vehicle_params.path_selection_strategy;
-
-        if (!predicted_object.kinematics.predicted_paths.empty()) {
-          auto & paths = modified_predicted_object.kinematics.predicted_paths;
-          if (path_selection_strategy == "random") {
-            // Randomly select a path index
-            const size_t num_paths = predicted_object.kinematics.predicted_paths.size();
-            std::uniform_int_distribution<size_t> path_index_dist(0, num_paths - 1);
-            const size_t random_path_index = path_index_dist(random_generator_);
-            // Reorder paths to put the randomly selected path first
-            std::swap(paths[0], paths[random_path_index]);
-
-            RCLCPP_DEBUG(
-              rclcpp::get_logger("dummy_perception_publisher"),
-              "Randomly selected path %zu out of %zu for %s object %s", random_path_index,
-              num_paths, is_pedestrian ? "pedestrian" : "vehicle", obj_uuid_str.c_str());
-          } else if (path_selection_strategy == "highest_confidence") {
-            // Find path with highest confidence and move it to first position
-            auto max_confidence_it = std::max_element(
-              paths.begin(), paths.end(),
-              [](const auto & a, const auto & b) { return a.confidence < b.confidence; });
-
-            if (max_confidence_it != paths.begin()) {
-              std::swap(paths[0], *max_confidence_it);
-            }
-
-            RCLCPP_DEBUG(
-              rclcpp::get_logger("dummy_perception_publisher"),
-              "Selected most likely path (confidence: %.3f) for %s object %s", paths[0].confidence,
-              is_pedestrian ? "pedestrian" : "vehicle", obj_uuid_str.c_str());
-          }
-        }
-
-        // Store this as the new prediction to use for some seconds
-        dummy_predicted_info_map[obj_uuid_str].last_used_prediction = modified_predicted_object;
-        dummy_predicted_info_map[obj_uuid_str].last_used_prediction_time = msg_time;
-        dummy_predicted_info_map[obj_uuid_str].prediction_update_timestamp = current_time;
-
-        return std::make_pair(modified_predicted_object, msg_time);
-      }
-    }
-  }
-
-  return std::make_pair(empty_object, empty_time);
-}
-
-std::set<std::string> DummyPerceptionPublisherNode::collectAvailablePredictedUUIDs(
-  const PredictedObjects & predicted_objects, std::map<std::string, Point> & predicted_positions)
-{
-  std::set<std::string> available_predicted_uuids;
-
-  for (const auto & pred_obj : predicted_objects.objects) {
-    const auto pred_uuid_str = autoware_utils_uuid::to_hex_string(pred_obj.object_id);
-    available_predicted_uuids.insert(pred_uuid_str);
-    predicted_positions[pred_uuid_str] =
-      pred_obj.kinematics.initial_pose_with_covariance.pose.position;
-  }
-
-  return available_predicted_uuids;
-}
-
-std::vector<std::string> DummyPerceptionPublisherNode::findDisappearedPredictedObjectUUIDs(
-  std::set<std::string> & available_predicted_uuids)
-{
-  std::vector<std::string> dummy_objects_to_remap;
-  auto & dummy_predicted_info_map = predicted_dummy_objects_tracking_info_.dummy_predicted_info_map;
-  for (const auto & mapping : dummy_predicted_info_map) {
-    const std::string & dummy_uuid = mapping.first;
-    const std::string & predicted_uuid = mapping.second.predicted_uuid;
-
-    // Skip entries without a predicted UUID mapping
-    if (predicted_uuid.empty()) {
-      continue;
-    }
-
-    // If the predicted object ID no longer exists, mark dummy for remapping
-    if (available_predicted_uuids.find(predicted_uuid) == available_predicted_uuids.end()) {
-      dummy_objects_to_remap.push_back(dummy_uuid);
-    } else {
-      // Remove already assigned predicted objects from available set
-      available_predicted_uuids.erase(predicted_uuid);
-    }
-  }
-
-  return dummy_objects_to_remap;
-}
-
-std::map<std::string, Point> DummyPerceptionPublisherNode::collectDummyObjectPositions(
-  const std::vector<tier4_simulation_msgs::msg::DummyObject> & dummy_objects,
-  const rclcpp::Time & current_time, std::vector<std::string> & unmapped_dummy_uuids)
-{
-  std::map<std::string, Point> dummy_positions;
-  auto & dummy_predicted_info_map = predicted_dummy_objects_tracking_info_.dummy_predicted_info_map;
-
-  for (const auto & dummy_obj : dummy_objects) {
-    const auto dummy_uuid_str = autoware_utils_uuid::to_hex_string(dummy_obj.id);
-
-    // Use last known position if available (which includes straight-line calculated position)
-    // Otherwise calculate current position using straight-line model
-    auto info_it = dummy_predicted_info_map.find(dummy_uuid_str);
-
-    dummy_positions[dummy_uuid_str] =
-      (info_it != dummy_predicted_info_map.end() && info_it->second.last_known_position.has_value())
-        ? info_it->second.last_known_position.value()
-        : ObjectInfo::calculateStraightLinePosition(dummy_obj, current_time).position;
-
-    if (info_it == dummy_predicted_info_map.end() || info_it->second.predicted_uuid.empty()) {
-      unmapped_dummy_uuids.push_back(dummy_uuid_str);
-    }
-  }
-
-  return dummy_positions;
-}
-
-std::optional<std::string> DummyPerceptionPublisherNode::findBestPredictedObjectMatch(
-  const std::string & dummy_uuid, const Point & dummy_position,
-  const std::set<std::string> & available_predicted_uuids,
-  const std::map<std::string, Point> & predicted_positions,
-  const PredictedObjects & predicted_objects)
-{
-  // Get the best matching predicted object based on several metrics
-  std::string closest_pred_uuid;
-  double min_distance = std::numeric_limits<double>::max();
-
-  // Iterate over all available predicted objects to find the best match
-  for (const auto & pred_uuid : available_predicted_uuids) {
-    const auto & pred_pos = predicted_positions.at(pred_uuid);
-    auto pred_obj = std::find_if(
-      predicted_objects.objects.begin(), predicted_objects.objects.end(),
-      [&pred_uuid](const auto & obj) {
-        return autoware_utils_uuid::to_hex_string(obj.object_id) == pred_uuid;
-      });
-
-    // Handle case where predicted object is not found (should not happen)
-    if (pred_obj == predicted_objects.objects.end()) {
-      continue;
-    }
-
-    // Find the actual predicted object for validation
-    const PredictedObject & candidate_pred_obj = *pred_obj;
-
-    // In case of multiple valid candidates, choose the closest one
-    double distance = calculateEuclideanDistance(dummy_position, pred_pos);
-
-    // Check if there is a valid remapping candidate based on position and speed
-    if (
-      distance >= min_distance ||
-      !isValidRemappingCandidate(candidate_pred_obj, dummy_uuid, dummy_position)) {
-      continue;
-    }
-
-    min_distance = distance;
-    closest_pred_uuid = pred_uuid;
-  }
-
-  if (closest_pred_uuid.empty()) {
-    return std::nullopt;
-  }
-
-  return closest_pred_uuid;
-}
-
-void DummyPerceptionPublisherNode::createRemappingsForDisappearedObjects(
-  const std::vector<std::string> & dummy_objects_to_remap,
-  std::set<std::string> & available_predicted_uuids,
-  const std::map<std::string, Point> & predicted_positions,
-  const std::map<std::string, Point> & dummy_positions, const PredictedObjects & predicted_objects)
-{
-  const rclcpp::Time current_time = this->now();
-  auto & dummy_predicted_info_map = predicted_dummy_objects_tracking_info_.dummy_predicted_info_map;
-  // First, remove old mappings
-  for (const auto & dummy_uuid : dummy_objects_to_remap) {
-    auto it = dummy_predicted_info_map.find(dummy_uuid);
-    if (it != dummy_predicted_info_map.end()) {
-      it->second.predicted_uuid.clear();
-    }
-  }
-
-  // Find best matches for all objects that need remapping
-  std::vector<std::pair<std::string, double>> mapping_candidates;
-
-  for (const auto & dummy_uuid : dummy_objects_to_remap) {
-    // Use last known position if available, otherwise use current position
-    Point remapping_position;
-    auto current_pos_it = dummy_positions.find(dummy_uuid);
-    if (current_pos_it == dummy_positions.end()) {
-      continue;
-    }
-    auto info_it = dummy_predicted_info_map.find(dummy_uuid);
-
-    remapping_position =
-      (info_it != dummy_predicted_info_map.end() && info_it->second.last_known_position.has_value())
-        ? info_it->second.last_known_position.value()
-        : current_pos_it->second;
-    // Find closest available predicted object for remapping
-    auto best_match = findBestPredictedObjectMatch(
-      dummy_uuid, remapping_position, available_predicted_uuids, predicted_positions,
-      predicted_objects);
-
-    if (best_match) {
-      const double distance =
-        calculateEuclideanDistance(remapping_position, predicted_positions.at(*best_match));
-      mapping_candidates.emplace_back(dummy_uuid + ":" + *best_match, distance);
-    }
-  }
-
-  // Sort candidates by distance to ensure closest matches get priority
-  // This is because multiple dummy objects may have the same best match
-  std::sort(
-    mapping_candidates.begin(), mapping_candidates.end(),
-    [](const auto & a, const auto & b) { return a.second < b.second; });
-
-  // Create mappings in order of proximity, ensuring one-to-one mapping
-  for (const auto & candidate : mapping_candidates) {
-    const std::string & combined = candidate.first;
-    const size_t colon_pos = combined.find(':');
-    const std::string dummy_uuid = combined.substr(0, colon_pos);
-    const std::string predicted_uuid = combined.substr(colon_pos + 1);
-
-    // Only create mapping if predicted object is still available
-    if (available_predicted_uuids.find(predicted_uuid) != available_predicted_uuids.end()) {
-      dummy_predicted_info_map[dummy_uuid].predicted_uuid = predicted_uuid;
-      dummy_predicted_info_map[dummy_uuid].mapping_timestamp = current_time;
-      available_predicted_uuids.erase(predicted_uuid);
-    }
-  }
-}
-
-double DummyPerceptionPublisherNode::calculateEuclideanDistance(
-  const Point & pos1, const Point & pos2)
-{
-  double dx = pos1.x - pos2.x;
-  double dy = pos1.y - pos2.y;
-  double dz = pos1.z - pos2.z;
-  return std::sqrt(dx * dx + dy * dy + dz * dz);
-}
-
-bool DummyPerceptionPublisherNode::isValidRemappingCandidate(
-  const PredictedObject & candidate_prediction, const std::string & dummy_uuid_str,
-  const Point & expected_position)
-{
-  // Perform various checks to validate if the candidate predicted object is a good match
-  // for the dummy object
-  auto dummy_it =
-    std::find_if(objects_.begin(), objects_.end(), [&dummy_uuid_str](const auto & obj) {
-      const auto uuid = autoware_utils_uuid::to_hex_string(obj.id);
-      return uuid == dummy_uuid_str;
-    });
-
-  if (dummy_it == objects_.end()) {
-    return false;
-  }
-
-  const auto & dummy_object = *dummy_it;
-  // Check if this is a pedestrian object
-
-  bool is_pedestrian =
-    (dummy_object.classification.label ==
-     autoware_perception_msgs::msg::ObjectClassification::PEDESTRIAN);
-
-  // Use class-specific thresholds
-  // Pedestrians: more lenient due to unpredictable movement patterns
-  // Vehicles: stricter for more predictable movement
-  const auto & pedestrian_params = predicted_object_params_.pedestrian_params;
-  const auto & vehicle_params = predicted_object_params_.vehicle_params;
-  const double max_remapping_distance = is_pedestrian ? pedestrian_params.max_remapping_distance
-                                                      : vehicle_params.max_remapping_distance;
-  const double max_speed_difference_ratio = is_pedestrian
-                                              ? pedestrian_params.max_speed_difference_ratio
-                                              : vehicle_params.max_speed_difference_ratio;
-
-  // Check if candidate has predicted paths
-  if (candidate_prediction.kinematics.predicted_paths.empty()) {
-    return false;
-  }
-
-  // Get dummy object speed for comparison (reuse dummy_it from above)
-  const double dummy_speed = dummy_object.initial_state.twist_covariance.twist.linear.x;
-
-  // Get candidate predicted object speed
-  const auto & candidate_twist =
-    candidate_prediction.kinematics.initial_twist_with_covariance.twist;
-  const double candidate_speed = std::sqrt(
-    candidate_twist.linear.x * candidate_twist.linear.x +
-    candidate_twist.linear.y * candidate_twist.linear.y);
-
-  // Speed bounds check - more lenient for pedestrians
-  const double min_speed_ratio =
-    is_pedestrian ? pedestrian_params.min_speed_ratio : vehicle_params.min_speed_ratio;
-  const double max_speed_ratio =
-    is_pedestrian ? pedestrian_params.max_speed_ratio : vehicle_params.max_speed_ratio;
-  const double speed_check_threshold =
-    is_pedestrian ? pedestrian_params.speed_check_threshold : vehicle_params.speed_check_threshold;
-
-  auto is_within_speed_bounds = [&](double speed) {
-    return speed >= min_speed_ratio * dummy_speed && speed <= max_speed_ratio * dummy_speed;
-  };
-
-  // Reject if candidate speed is too different when dummy speed is significant
-  if (dummy_speed > speed_check_threshold && !is_within_speed_bounds(candidate_speed)) {
-    RCLCPP_DEBUG(
-      rclcpp::get_logger("dummy_perception_publisher"),
-      "Rejecting remapping candidate for object %s (%s) due to speed difference: dummy=%fm/s, "
-      "candidate=%fm/s",
-      dummy_uuid_str.c_str(), is_pedestrian ? "pedestrian" : "vehicle", dummy_speed,
-      candidate_speed);
-    return false;
-  }
-
-  // Compare speeds if both are significant
-  if (dummy_speed > 0.1 && candidate_speed > 0.1) {
-    const double speed_ratio =
-      std::max(dummy_speed / candidate_speed, candidate_speed / dummy_speed);
-    if (speed_ratio > max_speed_difference_ratio) {
-      RCLCPP_DEBUG(
-        rclcpp::get_logger("dummy_perception_publisher"),
-        "Rejecting remapping candidate for object %s (%s) due to speed difference: %fx (dummy: "
-        "%fm/s, "
-        "candidate: %fm/s, max_ratio: %f)",
-        dummy_uuid_str.c_str(), is_pedestrian ? "pedestrian" : "vehicle", speed_ratio, dummy_speed,
-        candidate_speed, max_speed_difference_ratio);
-      return false;
-    }
-  }
-
-  // Calculate expected position based on last known trajectory if available
-  Point comparison_position = expected_position;
-  auto & dummy_predicted_info_map = predicted_dummy_objects_tracking_info_.dummy_predicted_info_map;
-  auto info_it = dummy_predicted_info_map.find(dummy_uuid_str);
-  if (
-    info_it == dummy_predicted_info_map.end() ||
-    !info_it->second.last_used_prediction.has_value()) {
-    return true;  // No last prediction, allow the match
-  }
-
-  // Calculate where the dummy object should be based on its last known trajectory
-  const auto & last_trajectory =
-    info_it->second.last_used_prediction.value().kinematics.predicted_paths.at(0);
-  const auto expected_pos = calculateExpectedPosition(last_trajectory, dummy_uuid_str);
-  if (expected_pos.has_value()) {
-    comparison_position = expected_pos.value();
-  }
-
-  // Check position similarity
-  const auto & candidate_pos =
-    candidate_prediction.kinematics.initial_pose_with_covariance.pose.position;
-  const double distance = calculateEuclideanDistance(comparison_position, candidate_pos);
-
-  if (distance > max_remapping_distance) {
-    RCLCPP_DEBUG(
-      rclcpp::get_logger("dummy_perception_publisher"),
-      "Rejecting remapping candidate for object %s (%s) due to large distance: %fm > %fm "
-      "(expected: %f, %f, "
-      "candidate: %f, %f)",
-      dummy_uuid_str.c_str(), is_pedestrian ? "pedestrian" : "vehicle", distance,
-      max_remapping_distance, comparison_position.x, comparison_position.y, candidate_pos.x,
-      candidate_pos.y);
-    return false;
-  }
-
-  return true;
-}
-
-std::optional<Point> DummyPerceptionPublisherNode::calculateExpectedPosition(
-  const autoware_perception_msgs::msg::PredictedPath & last_prediction,
-  const std::string & dummy_uuid_str)
-{
-  // Check if we have predicted paths
-  if (last_prediction.path.empty()) {
-    return std::nullopt;
-  }
-
-  // Find the dummy object by UUID
-  auto dummy_it =
-    std::find_if(objects_.begin(), objects_.end(), [&dummy_uuid_str](const auto & obj) {
-      const auto uuid = autoware_utils_uuid::to_hex_string(obj.id);
-      return uuid == dummy_uuid_str;
-    });
-
-  if (dummy_it == objects_.end()) {
-    return std::nullopt;
-  }
-
-  const auto & dummy_object = *dummy_it;
-
-  const auto & selected_path = last_prediction;
-
-  if (selected_path.path.empty()) {
-    return std::nullopt;
-  }
-
-  // If only one point in path, return that point
-  if (selected_path.path.size() < 2) {
-    return selected_path.path.back().position;
-  }
-
-  // Calculate elapsed time from last prediction to current time
-  const auto current_time = get_clock()->now();
-  auto & dummy_predicted_info_map = predicted_dummy_objects_tracking_info_.dummy_predicted_info_map;
-  auto info_it = dummy_predicted_info_map.find(dummy_uuid_str);
-  if (
-    info_it == dummy_predicted_info_map.end() ||
-    !info_it->second.last_used_prediction_time.has_value()) {
-    return std::nullopt;
-  }
-
-  const double elapsed_time =
-    (current_time - info_it->second.last_used_prediction_time.value()).seconds();
-
-  // Calculate distance traveled based on elapsed time and dummy object speed
-  const double speed = dummy_object.initial_state.twist_covariance.twist.linear.x;
-  const double distance_traveled = speed * elapsed_time;
-
-  // Calculate cumulative distances along the path
-  std::vector<double> cumulative_distances;
-  cumulative_distances.reserve(selected_path.path.size());
-  cumulative_distances.push_back(0.0);
-
-  for (size_t i = 1; i < selected_path.path.size(); ++i) {
-    const auto & prev_pose = selected_path.path.at(i - 1);
-    const auto & curr_pose = selected_path.path.at(i);
-
-    const double dx = curr_pose.position.x - prev_pose.position.x;
-    const double dy = curr_pose.position.y - prev_pose.position.y;
-    const double dz = curr_pose.position.z - prev_pose.position.z;
-    const double segment_length = std::sqrt(dx * dx + dy * dy + dz * dz);
-
-    cumulative_distances.push_back(cumulative_distances.back() + segment_length);
-  }
-
-  Point expected_position;
-
-  if (distance_traveled <= 0.0) {
-    // No movement, use initial position
-    return std::nullopt;
-  }
-  if (distance_traveled >= cumulative_distances.back()) {
-    // Extrapolate beyond the path end
-    // Use the last two points to determine direction and extrapolate
-    const auto & second_last_pose = selected_path.path.at(selected_path.path.size() - 2);
-    const auto & last_pose = selected_path.path.back();
-
-    // Calculate direction vector from second-last to last pose
-    const double dx = last_pose.position.x - second_last_pose.position.x;
-    const double dy = last_pose.position.y - second_last_pose.position.y;
-    const double dz = last_pose.position.z - second_last_pose.position.z;
-    const double segment_length = std::sqrt(dx * dx + dy * dy + dz * dz);
-
-    if (segment_length < std::numeric_limits<double>::epsilon()) {
-      return last_pose.position;
-    }
-    // Normalize direction vector
-    const double dir_x = dx / segment_length;
-    const double dir_y = dy / segment_length;
-    const double dir_z = dz / segment_length;
-
-    // Extrapolate position
-    const double overshoot_distance = distance_traveled - cumulative_distances.back();
-    expected_position.x = last_pose.position.x + dir_x * overshoot_distance;
-    expected_position.y = last_pose.position.y + dir_y * overshoot_distance;
-    expected_position.z = last_pose.position.z + dir_z * overshoot_distance;
-    return expected_position;
-  }
-
-  // Interpolate along the path
-  for (size_t i = 1; i < cumulative_distances.size(); ++i) {
-    if (distance_traveled > cumulative_distances.at(i)) {
-      continue;
-    }
-    // Interpolate between path points i-1 and i
-    const double segment_start_distance = cumulative_distances.at(i - 1);
-    const double segment_end_distance = cumulative_distances.at(i);
-    const double segment_length = segment_end_distance - segment_start_distance;
-
-    if (segment_length < std::numeric_limits<double>::epsilon()) {
-      return selected_path.path.at(i - 1).position;
-    }
-    const double interpolation_factor =
-      (distance_traveled - segment_start_distance) / segment_length;
-
-    const auto & start_pose = selected_path.path.at(i - 1);
-    const auto & end_pose = selected_path.path.at(1);
-
-    expected_position.x =
-      start_pose.position.x + interpolation_factor * (end_pose.position.x - start_pose.position.x);
-    expected_position.y =
-      start_pose.position.y + interpolation_factor * (end_pose.position.y - start_pose.position.y);
-    expected_position.z =
-      start_pose.position.z + interpolation_factor * (end_pose.position.z - start_pose.position.z);
-    return expected_position;
-  }
-  return selected_path.path.back().position;
-}
 }  // namespace autoware::dummy_perception_publisher
